+++ package.json
@@ -7,13 +7,19 @@
   "description": "A simple Node app built on Express, instantly up and running.",
   "main": "server.js",
   "scripts": {
-    "start": "node server.js"
+    "start": "node ./dist/server.js",
+    "build": "npm i typescript && npx tsc"
   },
   "dependencies": {
+    "@types/cors": "^2.8.13",
+    "@types/dotenv": "^8.2.0",
+    "@types/express": "^4.17.17",
+    "@types/lodash": "^4.14.191",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.17.1",
-    "lodash": "^4.17.21"
+    "lodash": "^4.17.21",
+    "typescript": "^5.0.2"
   },
   "repository": {
     "url": "https://glitch.com/edit/#!/hello-express"
