+++ screenshots/example_react_quotes_app.png
