+++ tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "compilerOptions": {
+    "esModuleInterop": true,
+    "outDir": "./dist",
+    "resolveJsonModule":true,
+  }
+}
